// Web/ClientApp/services/generated/polygon-area-model.ts
//----------------------
// <auto-generated>
//     Generated by Ozytis Model Generator  DO NOT EDIT!
// </auto-generated>
//----------------------

import { RestrictedAreaModel } from "./restricted-area-model";
import { CoordinateModel } from "./coordinate-model";
export interface PolygonAreaModel extends RestrictedAreaModel
{
    polygon: CoordinateModel[];
}